demos/reverse-image-search/requirements.txt
pymysql==1.0.2
diskcache==4.1.0
uvicorn==0.13.4
numpy==1.19.5
pydantic==1.8.2
pymilvus==1.1.2
fastapi==0.65.2
python-multipart
aiofiles
pillow==8.3.1
towhee==0.4.0
